+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/FilterTableCellEditor.java
@@ -18,16 +18,20 @@
  */
 package is.codion.swing.common.ui.component.table;
 
+import is.codion.swing.common.ui.component.table.FilterTableCellEditor.Builder.ComponentStep;
 import is.codion.swing.common.ui.component.value.ComponentValue;
 
 import javax.swing.JComponent;
 import javax.swing.table.TableCellEditor;
+import java.util.EventObject;
 import java.util.Optional;
+import java.util.function.Function;
 import java.util.function.Supplier;
 
 /**
  * A {@link TableCellEditor} implementation for {@link FilterTable}.
  * @param <T> the value type
+ * @see #builder()
  */
 public interface FilterTableCellEditor<T> extends TableCellEditor {
 
@@ -37,13 +41,10 @@ public interface FilterTableCellEditor<T> extends TableCellEditor {
 	ComponentValue<T, ? extends JComponent> componentValue();
 
 	/**
-	 * Creates a new default {@link FilterTableCellEditor} instance.
-	 * @param inputComponent supplies the input component
-	 * @param <T> the cell value type
-	 * @return a new default {@link FilterTableCellEditor} instance
+	 * @return a {@link Builder.ComponentStep} instance.
 	 */
-	static <T> FilterTableCellEditor<T> filterTableCellEditor(Supplier<ComponentValue<T, ? extends JComponent>> inputComponent) {
-		return new DefaultFilterTableCellEditor<>(inputComponent);
+	static ComponentStep builder() {
+		return DefaultFilterTableCellEditor.DefaultBuilder.COMPONENT;
 	}
 
 	/**
@@ -56,5 +57,37 @@ public interface FilterTableCellEditor<T> extends TableCellEditor {
 		 * @return a {@link TableCellEditor} instance for the given column or an empty optional if the column should not be editable
 		 */
 		Optional<TableCellEditor> create(C identifier);
+	}
+
+	/**
+	 * @param <T> the cell type
+	 */
+	interface Builder<T> {
+
+		/**
+		 * Provides a {@link FilterTableCellEditor.Builder}
+		 */
+		interface ComponentStep {
+
+			/**
+			 * Creates a new default {@link FilterTableCellEditor.Builder} instance.
+			 * @param component supplies the input component
+			 * @param <T> the cell value type
+			 * @return a new {@link FilterTableCellEditor.Builder} instance
+			 */
+			<T> Builder<T> component(Supplier<ComponentValue<T, ? extends JComponent>> component);
+		}
+
+		/**
+		 * @param cellEditable a function specifying if a cell is editable
+		 * @return this builder
+		 * @see TableCellEditor#isCellEditable(EventObject)
+		 */
+		Builder<T> cellEditable(Function<EventObject, Boolean> cellEditable);
+
+		/**
+		 * @return a new {@link FilterTableCellEditor} based on this builder
+		 */
+		FilterTableCellEditor<T> build();
 	}
 }